First pass at general refactor for donate modal
