+++ src/modals/ApproveTransactionModal/styles.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const StyledApproveTransaction = styled.div`
+  padding: 22px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  & > div {
+    margin-top: 15px;
+  }
+  & > p {
+    margin-top: 32px;
+  }
+`;
